Handle managers with multiple inheritance
